requirements.txt
# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tensorflow>=2.13.0
deepface>=0.0.79
tf-keras
# Image processing
Pillow>=10.0.0

# Optional but recommended
scikit-learn>=1.3.0
scipy>=1.11.0

flask>=2.3.0
flask-socketio>=5.3.0

# Utilities
tqdm>=4.66.0
